+++ tests/xml/test_constant.py
@@ -1,0 +1,352 @@
+"""Unit tests for constants and value specifications."""
+
+# pylint: disable=missing-class-docstring, missing-function-docstring
+import os
+import sys
+import unittest
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '../../src')))
+#import autosar.xml.enumeration as ar_enum # noqa E402
+import autosar.xml.element as ar_element # noqa E402
+import autosar # noqa E402
+
+class TestTextValueSpecification(unittest.TestCase): # noqa D101
+
+    def test_read_write_empty(self): # noqa D102
+        element = ar_element.TextValueSpecification()
+        writer = autosar.xml.Writer()
+        xml = writer.write_str_elem(element)
+        self.assertEqual(xml, '<TEXT-VALUE-SPECIFICATION/>')
+        reader = autosar.xml.Reader()
+        elem: ar_element.TextValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.TextValueSpecification)
+
+    def test_read_write_label(self): # noqa D102
+        element = ar_element.TextValueSpecification(label="MyLabel")
+        writer = autosar.xml.Writer()
+        xml = '''<TEXT-VALUE-SPECIFICATION>
+  <SHORT-LABEL>MyLabel</SHORT-LABEL>
+</TEXT-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.TextValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.TextValueSpecification)
+        self.assertEqual(elem.label, "MyLabel")
+
+    def test_read_write_value(self): # noqa D102
+        element = ar_element.TextValueSpecification(value="MyValue")
+        writer = autosar.xml.Writer()
+        xml = '''<TEXT-VALUE-SPECIFICATION>
+  <VALUE>MyValue</VALUE>
+</TEXT-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.TextValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.TextValueSpecification)
+        self.assertEqual(elem.value, "MyValue")
+
+class TestNumericalValueSpecification(unittest.TestCase): # noqa D101
+
+    def test_read_write_empty(self): # noqa D102
+        element = ar_element.NumericalValueSpecification()
+        writer = autosar.xml.Writer()
+        xml = writer.write_str_elem(element)
+        self.assertEqual(xml, '<NUMERICAL-VALUE-SPECIFICATION/>')
+        reader = autosar.xml.Reader()
+        elem: ar_element.NumericalValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NumericalValueSpecification)
+
+    def test_read_write_int_value(self): # noqa D102
+        element = ar_element.NumericalValueSpecification(value=4)
+        writer = autosar.xml.Writer()
+        xml = '''<NUMERICAL-VALUE-SPECIFICATION>
+  <VALUE>4</VALUE>
+</NUMERICAL-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NumericalValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(elem.value, 4)
+
+    def test_read_write_float_value(self): # noqa D102
+        element = ar_element.NumericalValueSpecification(value=4 / 10)
+        writer = autosar.xml.Writer()
+        xml = '''<NUMERICAL-VALUE-SPECIFICATION>
+  <VALUE>0.4</VALUE>
+</NUMERICAL-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NumericalValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NumericalValueSpecification)
+        self.assertAlmostEqual(elem.value, 0.4)
+
+class TestNotAvailableValueSpecification(unittest.TestCase): # noqa D101
+
+    def test_read_write_empty(self): # noqa D102
+        element = ar_element.NotAvailableValueSpecification()
+        writer = autosar.xml.Writer()
+        xml = writer.write_str_elem(element)
+        self.assertEqual(xml, '<NOT-AVAILABLE-VALUE-SPECIFICATION/>')
+        reader = autosar.xml.Reader()
+        elem: ar_element.NotAvailableValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NotAvailableValueSpecification)
+
+    def test_read_write_default_pattern(self): # noqa D102
+        element = ar_element.NotAvailableValueSpecification(default_pattern=255)
+        writer = autosar.xml.Writer()
+        xml = '''<NOT-AVAILABLE-VALUE-SPECIFICATION>
+  <DEFAULT-PATTERN>255</DEFAULT-PATTERN>
+</NOT-AVAILABLE-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.NotAvailableValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.NotAvailableValueSpecification)
+        self.assertEqual(elem.default_pattern, 255)
+
+
+class TestArrayValueSpecification(unittest.TestCase): # noqa D101
+
+    def test_read_write_empty(self): # noqa D102
+        element = ar_element.ArrayValueSpecification()
+        writer = autosar.xml.Writer()
+        xml = writer.write_str_elem(element)
+        self.assertEqual(xml, '<ARRAY-VALUE-SPECIFICATION/>')
+        reader = autosar.xml.Reader()
+        elem: ar_element.ArrayValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ArrayValueSpecification)
+
+    def test_read_write_label(self): # noqa D102
+        element = ar_element.ArrayValueSpecification(label="MyLabel")
+        writer = autosar.xml.Writer()
+        xml = '''<ARRAY-VALUE-SPECIFICATION>
+  <SHORT-LABEL>MyLabel</SHORT-LABEL>
+</ARRAY-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ArrayValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(elem.label, "MyLabel")
+
+    def test_read_write_numerical_elements(self): # noqa D102
+        element = ar_element.ArrayValueSpecification()
+        element.append(ar_element.NumericalValueSpecification(value=1))
+        element.append(ar_element.NumericalValueSpecification(value=2))
+        element.append(ar_element.NumericalValueSpecification(value=3))
+        writer = autosar.xml.Writer()
+        xml = '''<ARRAY-VALUE-SPECIFICATION>
+  <ELEMENTS>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <VALUE>1</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <VALUE>2</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <VALUE>3</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+  </ELEMENTS>
+</ARRAY-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ArrayValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(elem.elements), 3)
+        child_elem = elem.elements[0]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.value, 1)
+        child_elem = elem.elements[1]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.value, 2)
+        child_elem = elem.elements[2]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.value, 3)
+
+    def test_read_write_text_elements(self): # noqa D102
+        element = ar_element.ArrayValueSpecification()
+        element.append(ar_element.TextValueSpecification(value="First"))
+        element.append(ar_element.TextValueSpecification(value="Second"))
+        writer = autosar.xml.Writer()
+        xml = '''<ARRAY-VALUE-SPECIFICATION>
+  <ELEMENTS>
+    <TEXT-VALUE-SPECIFICATION>
+      <VALUE>First</VALUE>
+    </TEXT-VALUE-SPECIFICATION>
+    <TEXT-VALUE-SPECIFICATION>
+      <VALUE>Second</VALUE>
+    </TEXT-VALUE-SPECIFICATION>
+  </ELEMENTS>
+</ARRAY-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.ArrayValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(elem.elements), 2)
+        child_elem = elem.elements[0]
+        self.assertIsInstance(child_elem, ar_element.TextValueSpecification)
+        self.assertEqual(child_elem.value, "First")
+        child_elem = elem.elements[1]
+        self.assertIsInstance(child_elem, ar_element.TextValueSpecification)
+        self.assertEqual(child_elem.value, "Second")
+
+class TestRecordValueSpecification(unittest.TestCase): # noqa D101
+
+    def test_read_write_empty(self): # noqa D102
+        element = ar_element.RecordValueSpecification()
+        writer = autosar.xml.Writer()
+        xml = writer.write_str_elem(element)
+        self.assertEqual(xml, '<RECORD-VALUE-SPECIFICATION/>')
+        reader = autosar.xml.Reader()
+        elem: ar_element.RecordValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.RecordValueSpecification)
+
+    def test_read_write_numerical_elements(self): # noqa D102
+        element = ar_element.RecordValueSpecification()
+        element.append(ar_element.NumericalValueSpecification(label="First", value=1))
+        element.append(ar_element.NumericalValueSpecification(label="Second", value=2))
+        element.append(ar_element.NumericalValueSpecification(label="Third", value=3))
+        writer = autosar.xml.Writer()
+        xml = '''<RECORD-VALUE-SPECIFICATION>
+  <FIELDS>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <SHORT-LABEL>First</SHORT-LABEL>
+      <VALUE>1</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <SHORT-LABEL>Second</SHORT-LABEL>
+      <VALUE>2</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+    <NUMERICAL-VALUE-SPECIFICATION>
+      <SHORT-LABEL>Third</SHORT-LABEL>
+      <VALUE>3</VALUE>
+    </NUMERICAL-VALUE-SPECIFICATION>
+  </FIELDS>
+</RECORD-VALUE-SPECIFICATION>'''
+        self.assertEqual(writer.write_str_elem(element), xml)
+        reader = autosar.xml.Reader()
+        elem: ar_element.RecordValueSpecification = reader.read_str_elem(xml)
+        self.assertIsInstance(elem, ar_element.RecordValueSpecification)
+        self.assertEqual(len(elem.fields), 3)
+        child_elem = elem.fields[0]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.label, "First")
+        self.assertEqual(child_elem.value, 1)
+        child_elem = elem.fields[1]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.label, "Second")
+        self.assertEqual(child_elem.value, 2)
+        child_elem = elem.fields[2]
+        self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+        self.assertEqual(child_elem.label, "Third")
+        self.assertEqual(child_elem.value, 3)
+
+
+class TestValueBuilder(unittest.TestCase): # noqa D101
+
+    def test_make_numerical_from_int(self): # noqa D102
+        element: ar_element.NumericalValueSpecification = ar_element.ValueSpecification.make_value(10)
+        self.assertIsInstance(element, ar_element.NumericalValueSpecification)
+        self.assertIsNone(element.label)
+        self.assertEqual(element.value, 10)
+
+        element: ar_element.NumericalValueSpecification = ar_element.ValueSpecification.make_value(("MyLabel", 20))
+        self.assertIsInstance(element, ar_element.NumericalValueSpecification)
+        self.assertEqual(element.label, "MyLabel")
+        self.assertEqual(element.value, 20)
+
+    def test_make_text_str(self):  # noqa D102
+        element: ar_element.TextValueSpecification = ar_element.ValueSpecification.make_value("Value")
+        self.assertIsInstance(element, ar_element.TextValueSpecification)
+        self.assertIsNone(element.label)
+        self.assertEqual(element.value, "Value")
+
+
+    def test_make_not_available_from_none(self):  # noqa D102
+        element: ar_element.NotAvailableValueSpecification = ar_element.ValueSpecification.make_value(None)
+        self.assertIsInstance(element, ar_element.NotAvailableValueSpecification)
+        self.assertIsNone(element.label)
+
+        element: ar_element.NotAvailableValueSpecification
+        element = ar_element.ValueSpecification.make_value(("MyLabel", None, 255))
+        self.assertIsInstance(element, ar_element.NotAvailableValueSpecification)
+        self.assertEqual(element.label, "MyLabel")
+        self.assertEqual(element.default_pattern, 255)
+
+    def test_make_array_from_list(self):  # noqa D102
+        element: ar_element.ArrayValueSpecification
+        element = ar_element.ValueSpecification.make_value(["ARRAY", 1, 2, 3])
+        self.assertIsInstance(element, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(element.elements), 3)
+        for i, child_elem in enumerate(element.elements, 1):
+            self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+            self.assertEqual(child_elem.value, i)
+
+    def test_make_multi_level_array(self):  # noqa D102
+        element: ar_element.ArrayValueSpecification
+        element = ar_element.ValueSpecification.make_value(["ARRAY",
+                                                            ["ARRAY", 1, 2, 3],
+                                                            ["ARRAY", 4, 5, 6],
+                                                            ["ARRAY", 7, 8, 9]
+                                                            ])
+        self.assertIsInstance(element, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(element.elements), 3)
+        child_elem: ar_element.ArrayValueSpecification = element.elements[0]
+        self.assertIsInstance(child_elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(child_elem.elements), 3)
+        for i, grand_child in enumerate(child_elem.elements, 1):
+            self.assertIsInstance(grand_child, ar_element.NumericalValueSpecification)
+            self.assertEqual(grand_child.value, i)
+        child_elem = element.elements[1]
+        self.assertIsInstance(child_elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(child_elem.elements), 3)
+        for i, grand_child in enumerate(child_elem.elements, 4):
+            self.assertIsInstance(grand_child, ar_element.NumericalValueSpecification)
+            self.assertEqual(grand_child.value, i)
+        child_elem = element.elements[2]
+        self.assertIsInstance(child_elem, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(child_elem.elements), 3)
+        for i, grand_child in enumerate(child_elem.elements, 7):
+            self.assertIsInstance(grand_child, ar_element.NumericalValueSpecification)
+            self.assertEqual(grand_child.value, i)
+
+    def test_make_record_from_list(self):
+        element: ar_element.RecordValueSpecification
+        element = ar_element.ValueSpecification.make_value(["RECORD", 1, 2, 3])
+        self.assertIsInstance(element, ar_element.RecordValueSpecification)
+        self.assertEqual(len(element.fields), 3)
+        for i, child_elem in enumerate(element.fields, 1):
+            self.assertIsInstance(child_elem, ar_element.NumericalValueSpecification)
+            self.assertEqual(child_elem.value, i)
+
+    def test_make_array_of_records_with_labels(self):
+        element: ar_element.ArrayValueSpecification
+        element = ar_element.ValueSpecification.make_value(["ARRAY",
+                                                            ["RECORD",
+                                                             ("First", 1),
+                                                             ("Second", 2),
+                                                             ("Third", 3),
+                                                             ],
+                                                            ["RECORD",
+                                                             ("First", 4),
+                                                             ("Second", 5),
+                                                             ("Third", 6),
+                                                             ],
+                                                            ["RECORD",
+                                                             ("First", 7),
+                                                             ("Second", 8),
+                                                             ("Third", 9),
+                                                             ],
+                                                            ])
+        self.assertIsInstance(element, ar_element.ArrayValueSpecification)
+        self.assertEqual(len(element.elements), 3)
+        child_elem: ar_element.RecordValueSpecification
+        labels = ["First", "Second", "Third"]
+        for i, child_elem in enumerate(element.elements, 0):
+            self.assertIsInstance(child_elem, ar_element.RecordValueSpecification)
+            self.assertEqual(len(child_elem.fields), 3)
+            grand_child: ar_element.NumericalValueSpecification
+            for j, grand_child in enumerate(child_elem.fields):
+                self.assertEqual(grand_child.label, labels[j])
+                self.assertEqual(grand_child.value, i * 3 + j + 1)
+
+
+if __name__ == '__main__':
+    unittest.main()